Catter
Alright devs, what we're going to do is we're going to make the chat functional. You see, the chat works, but you can't see others' messages, and they can't see yours. What I want the devs to do is to replace SharkyShar's cloud chat engine with a working one (such as the one used in Servo).